// src/app/core/models/laravel-pagination.ts
export interface LaravelPagination<T> {
    data: T[];
    links: {
        first: string;
        last: string;
        prev: string | null;
        next: string | null;
    };
    meta: {
        current_page: number;
        last_page: number;
        per_page: number;
        total: number;
        from?: number;
        to?: number;
        path?: string;
        links?: any[];
    };
}